requirements.txt
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
click==8.1.8
decorator==5.2.1
fastapi==0.115.12
flask==3.1.0
h11==0.14.0
idna==3.10
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLalchemy==2.0.40
starlette==0.46.2
typing_extensions==4.13.2
uvicorn==0.34.1
validators==0.34.0